final draft project/requirements.txt
datetime
holoviews
hvplot.pandas
numpy
os
pandas
panel
pytz
sklearn.cluster
sklearn.preprocessing 
yfinance
